requirements/requirements-gpu.txt
pandas==0.22.0
numpy==1.13.1
tensorflow-gpu==1.5.0
